+++ src/main/java/roomescape/core/service/ReservationTimeService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.core.domain.ReservationTime;
+import roomescape.core.dto.BookingTimeResponseDto;
 import roomescape.core.dto.ReservationTimeRequestDto;
 import roomescape.core.dto.ReservationTimeResponseDto;
 import roomescape.core.repository.ReservationRepository;
@@ -34,6 +35,11 @@ public class ReservationTimeService {
                 .stream()
                 .map(ReservationTimeResponseDto::new)
                 .toList();
+    }
+
+    @Transactional(readOnly = true)
+    public List<BookingTimeResponseDto> findBookable(final String date, final long themeId) {
+        return reservationRepository.findAllByDateAndThemeId(date, themeId);
     }
 
     @Transactional
