+++ src/main/java/roomescape/web/controller/ReservationController.java
@@ -10,9 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.core.dto.BookingTimeResponseDto;
 import roomescape.core.dto.ReservationRequestDto;
 import roomescape.core.dto.ReservationResponseDto;
 import roomescape.core.service.ReservationService;
@@ -36,12 +34,6 @@ public class ReservationController {
     @GetMapping
     public ResponseEntity<List<ReservationResponseDto>> findAll() {
         return ResponseEntity.ok(reservationService.findAll());
-    }
-
-    @GetMapping(params = {"date", "themeId"})
-    public ResponseEntity<List<BookingTimeResponseDto>> findBookable(@RequestParam("date") String date,
-                                                                     @RequestParam("themeId") Long themeId) {
-        return ResponseEntity.ok(reservationService.findBookable(date, themeId));
     }
 
     @DeleteMapping("/{id}")
