+++ src/test/java/roomescape/web/controller/ReservationControllerTest.java
@@ -9,7 +9,6 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -22,7 +21,6 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
-import roomescape.core.dto.BookingTimeResponseDto;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
@@ -257,39 +255,6 @@ class ReservationControllerTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(countReservation()));
-    }
-
-    @Test
-    @DisplayName("날짜와 테마 정보가 주어지면 예약 가능한 시간 목록을 조회한다.")
-    void findBookable() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", TOMORROW_DATE);
-        params.put("timeId", 1);
-        params.put("themeId", 1);
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(201);
-
-        List<BookingTimeResponseDto> times = RestAssured.given().log().all()
-                .when().get("/reservations?date=" + TOMORROW_DATE + "&themeId=1")
-                .then().log().all()
-                .statusCode(200).extract()
-                .jsonPath().getList(".", BookingTimeResponseDto.class);
-
-        assertThat(times).hasSize(countReservationTime())
-                .allMatch(response -> validateBookingTime(response, 1));
-    }
-
-    private boolean validateBookingTime(final BookingTimeResponseDto bookingTime, final int timeId) {
-        if (bookingTime.getId() == timeId) {
-            return bookingTime.isAlreadyBooked();
-        }
-        return !bookingTime.isAlreadyBooked();
     }
 
     @Test
