+++ src/test/java/roomescape/web/controller/ReservationTimeControllerTest.java
@@ -1,10 +1,14 @@
 package roomescape.web.controller;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.is;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -17,11 +21,16 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
+import roomescape.core.dto.BookingTimeResponseDto;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @TestPropertySource(properties = {"spring.config.location = classpath:application-test.yml"})
 class ReservationTimeControllerTest {
+    private static final String TOMORROW_DATE = LocalDate.now()
+            .plusDays(1)
+            .format(DateTimeFormatter.ISO_DATE);
+
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
@@ -91,6 +100,40 @@ class ReservationTimeControllerTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(countReservationTime()));
+    }
+
+
+    @Test
+    @DisplayName("날짜와 테마 정보가 주어지면 예약 가능한 시간 목록을 조회한다.")
+    void findBookable() {
+        Map<String, Object> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", TOMORROW_DATE);
+        params.put("timeId", 1);
+        params.put("themeId", 1);
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(201);
+
+        List<BookingTimeResponseDto> times = RestAssured.given().log().all()
+                .when().get("/times?date=" + TOMORROW_DATE + "&themeId=1")
+                .then().log().all()
+                .statusCode(200).extract()
+                .jsonPath().getList(".", BookingTimeResponseDto.class);
+
+        assertThat(times).hasSize(countReservationTime())
+                .allMatch(response -> validateBookingTime(response, 1));
+    }
+
+    private boolean validateBookingTime(final BookingTimeResponseDto bookingTime, final int timeId) {
+        if (bookingTime.getId() == timeId) {
+            return bookingTime.isAlreadyBooked();
+        }
+        return !bookingTime.isAlreadyBooked();
     }
 
     @Test
